Update to Eclipse RDF4J Sesame
Mostly namespaces changes as per directions on
http://docs.rdf4j.org/migration/. Other changes are mostly URI-IRI and
Graph-Model related (deprecations).

Two tests still fail. Also note, version number updated because it was
strange to see Pinto depend on a version that does not exist on Github
(4.0.2).

It was also hard to make a decision as to how to name the dependency.
Chose to keep 'sesame' and where needed wrote 'sesameRdf4J' to clarify.